Added decimal and backspace support. Replaced int numbers with OperandString class. Performing operations with BigDecimal. Fixed issues with calculating based on result of a previous calculation.

+++ app/src/main/java/com/jacstuff/simplecalculator/actions/operators/Operator.java
@@ -1,8 +1,10 @@
 package com.jacstuff.simplecalculator.actions.operators;
 
+import java.math.BigDecimal;
+
 public interface Operator {
 
-    int execute(int number1, int number2);
+    BigDecimal execute(BigDecimal number1, BigDecimal number2);
     String getSymbol();
     void onLoad();
     boolean hasSingleInput();
